Update controllers for staff role

+++ backend-api/src/controllers/menu_items.controller.js
@@ -32,6 +32,17 @@ async function getItemsByFilter(req, res, next) {
 }
 
 async function deleteAllItems(req, res, next) {
+    if (!req.session.user) {
+        return next(new ApiError(401, 'Vui lòng đăng nhập để xem thông tin của bạn!'));
+    }
+
+    const userId = req.session.user.userid;
+    console.log("staffid: ", userId)
+    const userRole = await usersService.checkRole(userId);
+    
+    if (userRole !== 2) {
+        return next(new ApiError(403, 'Forbidden: Bạn không có quyền chỉnh sửa thông tin này!'));
+    }
     try{
         await menu_itemsService.deleteAllItems();
 
@@ -76,9 +87,8 @@ async function createItem(req, res, next) {
 
         // Kiểm tra vai trò người dùng
         const userRole = await usersService.checkRole(userId);
-        if (userRole !== 2  ) {
-            return next(new ApiError(403, 'Forbidden: You do not have permission to add menu items', { code: 'FORBIDDEN' }));
-        }
+        if (userRole == 2  ) {
+            
 
         if (!req.body?.item_name || typeof req.body.item_name !== 'string') {
             return next(new ApiError(400, 'Invalid input', { code: 'INVALID_INPUT' }));
@@ -104,7 +114,11 @@ async function createItem(req, res, next) {
             JSend.success({ 
               item
             })
-          );
+          );}
+          else{
+            return next(new ApiError(403, 'Forbidden: You do not have permission to add menu items', { code: 'FORBIDDEN' }));
+        
+          }
     } catch (error) {
         console.error(error);
         return next(new ApiError(500, 'An error occurred while creating the new menu item'));
@@ -201,85 +215,6 @@ async function deleteItem(req, res, next) {
     }
 }
 
-async function getTypeItemsByFilter(req, res, next) {
-    let result = {
-        items: [],
-        metadata: {
-            totalRecords: 0,
-            firstPage: 1,
-            lastPage: 1,
-            page: 1,
-            limit: 5,
-        }
-    };
-
-    try {
-        // Pass query parameters (e.g., item_name, item_type, item_status) for filtering
-        result = await menu_itemsService.getManyItemsByType(req.query);
-    } catch (error) {
-        console.error(error);
-        return next(new ApiError(500, 'An error occurred while retrieving menu items'));
-    }
-
-    return res.json(
-        JSend.success({
-            items: result.items,
-            metadata: result.metadata,
-        })
-    );
-}
-
-async function getManyMenuItemsByPrice(req, res, next) {
-    try {
-        // Extract query parameters, including min_price and max_price
-        const { min_price, max_price, page, limit } = req.query;
-
-        // Call the menu_itemsService function to get filtered menu items
-        const result = await menu_itemsService.getManyMenuItemsByPrice({
-            min_price,
-            max_price,
-            page,
-            limit
-        });
-
-        return res.json(JSend.success({ 
-            menu_items: result.menu_items,
-            metadata: result.metadata
-        }));
-    } catch (error) {
-        console.error(error);
-        return next(new ApiError(500, 'An error occurred while retrieving menu items'));
-    }
-}
-
-async function getManyMenuItemsByType_Price(req, res, next) {
-    try {
-        // Extract query parameters, including min_price and max_price
-        const { item_name, item_type, item_status, min_price, max_price, page, limit } = req.query;
-
-        // Call the menu_itemsService function to get filtered menu items
-        const result = await menu_itemsService.getManyMenuItemsByType_Price({
-            item_name,
-            item_type,
-            item_status,
-            min_price,
-            max_price,
-            page,
-            limit
-        });
-
-        return res.json(JSend.success({ 
-            menu_items: result.menu_items,
-            metadata: result.metadata
-        }));
-    } catch (error) {
-        console.error(error);
-        return next(new ApiError(500, 'An error occurred while retrieving menu items'));
-    }
-}
-
-
-
 async function getItemByName(req, res, next) {
   const {name } = req.params;
 
@@ -295,8 +230,6 @@ async function getItemByName(req, res, next) {
   }
 }
 
-
-
 module.exports = {
     getItemsByFilter,
     deleteAllItems,
@@ -304,10 +237,5 @@ module.exports = {
     createItem,
     updateItem,
     deleteItem,
-
-    getTypeItemsByFilter,
-    getManyMenuItemsByPrice,
-    getManyMenuItemsByType_Price,
     getItemByName,
-    
 };
